feat: 디테일 모달 html 생성 완료
